+++ package.json
@@ -13,7 +13,8 @@
     "react-helmet": "^6.1.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
-    "reactstrap": "^8.5.1"
+    "reactstrap": "^8.5.1",
+    "redux": "^4.0.5"
   },
   "scripts": {
     "start": "react-scripts start",
