--- src/Component/Cart/EmptyCart.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-class EmptyCart extends React.Component {
-  render() {
-    return <div className="empty-cart"></div>;
-  }
-}
-
-export default EmptyCart;
