+++ src/Component/CartItem/style.css
@@ -1,0 +1,25 @@
+/* Table - Product Image  */
+.product-name-img {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.ordered-product-name {
+  margin-right: 30px;
+}
+
+/* Table - number of item */
+.change-numberof-item-ip {
+  width: 30px;
+  text-align: center;
+  margin: 0px 3px;
+}
+
+.remove-item-btn {
+  background-color: #e23737;
+  color: #fff;
+  border: none;
+  margin-left: 10px;
+  border-radius: 5px;
+}
