+++ src/Component/Form/button/index.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+function Button(props) {
+  const { disabled } = props;
+  return (
+    <div className="btn-submit">
+      <button type="submit" disabled={disabled}>
+        Submit
+      </button>
+    </div>
+  );
+}
+
+export default Button;
