+++ src/Container/Signup/style.css
@@ -1,0 +1,98 @@
+/* Login page */
+.login-page,
+.signup-page {
+  /* 100% screen - height of header and footer */
+  min-height: calc(100vh - 90px - 66px);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: aliceblue;
+}
+/* Form */
+.form {
+  background-color: white;
+  width: 300px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  -webkit-box-shadow: 0 0 10px #fff;
+  box-shadow: 0 0 10px #fff;
+}
+
+.form-container {
+  width: 84%;
+  height: 90%;
+}
+/* Title */
+.title {
+  text-align: center;
+}
+
+.title > p {
+  margin: 18px 0px;
+  font-size: 30px;
+  font-weight: bold;
+}
+/* Input */
+.input-pw,
+.input-email {
+  margin: 0px 0px 20px 0px;
+}
+
+.input-pw > input,
+.input-email > input {
+  border: none;
+  border-radius: 1px;
+  box-sizing: border-box;
+  background-color: #efefef;
+  padding: 0px 0px 0px 10px;
+  width: 100%;
+  margin-bottom: 3px;
+  height: 40px;
+  font-size: 1.4em;
+}
+/* Span */
+.span-form {
+  display: inline-block;
+  font-size: 1.3em;
+  margin-bottom: 5px;
+}
+
+/* button */
+.btn-submit button {
+  width: 100%;
+  height: 40px;
+  color: white;
+  font-size: 1.3em;
+  background-color: var(--primary-color);
+  border: none;
+  outline: none;
+  margin-bottom: 14px;
+}
+
+/* Error */
+.error {
+  font-size: 12px;
+  color: red;
+  margin-bottom: 10px;
+}
+
+.has-not-account {
+  margin: 10px 0px 15px;
+  color: rgb(49, 43, 131);
+  text-decoration: none;
+  font-size: 12px;
+}
+
+.err-login {
+  color: red;
+  font-size: 12px;
+  margin-bottom: 15px;
+}
+
+/* forgot password*/
+.forgot-pw {
+  display: block;
+  color: #3f00ff;
+  margin-bottom: 10px;
+}
